Move 'CoreDataFeedStore' to its own production file Move Core Data related files into new Infrastructure/Core Data folder
