+++ FeedStoreChallenge/Infrastructure/Core Data/ManagedCache.swift
@@ -9,7 +9,6 @@
 
 import Foundation
 import CoreData
-import FeedStoreChallenge
 
 @objc(ManagedCache)
 public class ManagedCache: NSManagedObject {
