+++ FeedStoreChallenge/Infrastructure/Core Data/ManagedFeedImage.swift
@@ -9,7 +9,6 @@
 
 import Foundation
 import CoreData
-import FeedStoreChallenge
 
 @objc(ManagedFeedImage)
 public class ManagedFeedImage: NSManagedObject {
